+++ R/01-setModuleTitle.R
@@ -4,5 +4,7 @@
 #' @param titleTag (character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
 #'  \code{htmltools}
 setModuleTitle <- function(title, titleTag = "h4") {
+  if (is.null(title)) return(NULL)
+
   do.call(titleTag, list(title))
 }